WIP remove "highway" term and refactor code
