+++ pgdrive/tests/test_top_down_rendering/vis_highway_render.py
@@ -1,19 +1,18 @@
 import time
 
-from pgdrive.envs.pgdrive_env import PGDriveEnv
+from pgdrive.envs.top_down_env import TopDownPGDriveEnv
 from pgdrive.utils import setup_logger
 
 
 def vis_highway_render_with_panda_render():
     setup_logger(True)
 
-    env = PGDriveEnv(
+    env = TopDownPGDriveEnv(
         {
             "environment_num": 1,
             "manual_control": True,
             "use_render": True,
             "use_image": False,
-            "use_topdown": True,
             "traffic_mode": "reborn"
         }
     )
